Module worldgen::noise

The Noise module provides generators for different kinds of noise.

There are currently two different sources for noise: coherent and perlin. The coherent noise source provides no customisation and is very simple, and mainly exists to be used by the perlin source, which is the recommended one to use at the moment.

These generators provide a method for generating a noise value at a specific location, however are best used in combination with a NoiseMap

Modules

coherent
A provider of coherent noise. This generates a noise value for each corner of the unit square the given point is in, and then calculates the noise value based on these.

perlin
A provider of perlin noise.

Traits

NoiseProvider
The trait for a noise generator.