Struct worldgen::noisemap::NoiseMap

pub struct NoiseMap<T> {
    // some fields omitted
}

The standard noise map.

This is the base noise map, and is created by wrapping a noise source. It has properties that allow the setting of the generation seed, the size of the generated chunks, and the coordinate scale.

Methods

impl<T: NoiseProvider> NoiseMap<T>

fn new(noise: T) -> NoiseMap<T>

Construct a new noise map with the default properties.