Trait worldgen::noisemap::NoiseMapGenerator

pub trait NoiseMapGenerator: Clone + Copy {
    fn generate_chunk(&self, x: i32, y: i32) -> Vec<Vec<f64>>;
    fn set<P: Property>(self, property: P) -> Self;
    fn get_size(&self) -> Size;

    fn generate(&self) -> Vec<Vec<f64>> { ... }
}

Base trait for noise maps.

NoiseMap, ScaledNoiseMap, and NoiseMapCombination all implement this trait.

Required Methods

fn generate_chunk(&self, x: i32, y: i32) -> Vec<Vec<f64>>

Generates a specific chunk of the noise map.

This can be used to generate a larger map in smaller parts.

Example

for y in (0 .. 5) {
    for row in nm.generate_chunk(0, y).iter() {
        for value in row.iter() {
            print!("{}", value);
        }
        println!("");
    }
}

fn set<P: Property>(self, property: P) -> Self

Set a property on the noise map.

fn get_size(&self) -> Size

Returns the size of the noise map.

Provided Methods

fn generate(&self) -> Vec<Vec<f64>>

Generates the central chunk of the noise map.

The returned vector is a vector of rows of values.

This simply calls self.generate_chunk(0, 0).

Example

for row in nm.generate().iter() {
    for value in row.iter() {
        print!("{}", value);
    }
    println!("");
}