Module worldgen::world

For generating maps of specfic objects based on thresholds.

Once we have a NoiseMap we can use a World and generate a map of specific tiles using threshold constraints.

Reexports

pub use self::tile::Tile;

Modules

tile

Structs

Size
Sets the size of the generated chunks.

World
The World class.