Struct worldgen::world::World

pub struct World<NM, T> {
    // some fields omitted
}

The World class.

NM is the NoiseMap class, T is the type for each tile. See the module documentation for for information.

Methods

impl<NM: NoiseMapGenerator, T: Clone> World<NM, T>

fn new(nm: NM) -> World<NM, T>

Constructs a new world using a given noisemap

fn add(self, tile: Tile<T>) -> World<NM, T>

Add a tile definition to the world

fn generate(&self) -> Vec<Vec<T>>

Generates the central chunk of the world.

The returned vector is a vector of rows of tiles

This simply calls self.generate_chunk(0, 0).

Panics

This function will panic if a value is found which satisfies no tiles constraints.

Example

for row in world.generate().iter() {
    for tile in row.iter() {
        print!("{}", tile);
    }
    println!("");
}

fn generate_chunk(&self, x: i32, y: i32) -> Vec<Vec<T>>

Generates a specific chunk of the world.

Panics

This function will panic if a value is found which satisfies no tiles constraints.

Example

for y in (0 .. 5) {
    for row in world.generate_chunk(0, y).iter() {
        for tile in row.iter() {
            print!("{}", tile);
        }
        println!("");
    }
}