Module worldgen::world::tile

Structs

Tile
Objects to generate in the world based on given constraints

Enums

Constraint
A constraint that limits when a tile should be chosen for the generated world.