Struct worldgen::world::tile::Tile

pub struct Tile<T> {
    // some fields omitted
}

Objects to generate in the world based on given constraints

Methods

impl<T: Clone> Tile<T>

fn new(value: T) -> Tile<T>

Construct a new tile represented by the given object.

fn when(self, constraint: Constraint) -> Tile<T>

Adds a constraint to the tile.

fn value(&self) -> T

Returns the value this tile is represented by.

fn satisfied_by(&self, x: i64, y: i64, size: Size, chunk_x: i64, chunk_y: i64, nms: &mut HashMap<u64, Vec<Vec<f64>>>) -> bool

Returns true if the given value would satisfy all of this tile's constraints.