Module worldgen::world

For generating maps of specfic objects based on thresholds.

Once we have a NoiseMap we can use a World and generate a map of specific tiles using threshold constraints.

When you generate a world, it will generate the noisemap inside it, and then check each value against every tile definition, in order of their addition to the world.

The following example will produce the character ~ when the noise value is less than -0.7, the character ^ when the value is greater than 0.6, and , otherwise.

let world = World::new(nm)
    .add(Tile::new('~').when(Constraint::LT(-0.7)))
    .add(Tile::new('^').when(Constraint::GT(0.6)))
    .add(Tile::new(','));

Reexports

pub use self::tile::Tile;

Modules

tile

Structs

World
The World class.